Improve windows build compatibility (#243)
Add a make.bat roughly equivalent to the Makefile so that Windows users can more easily build the project. Also add a make winlint which ignores CRLF line endings as the default git on windows configuration will have CRLF line endings in the working tree but not in the repository itself.