Move warning about VEX/IFI's culture
While we by no means want to downplay the severity of these events, this
is a better spot for this warning as it is more relevant to selecting
vendors to buy from.

The direct link to the Chief Delphi thread is removed, because it
prevents the use of gm0 in some school environments.