Add notes about Actobotics being deprecated (#386)
This reflect Actobotics' status as a discontinued product.

This isn't necessarily a complete set of changes for Actobotics in gm0, there may be some places where more rewriting and refactoring is needed. My focus was to move it lower in the available lists and make it clearer up front that it's a now a discontinued line of products (rather than have readers have to figure that out on the ServoCity pages).