Use ghostscript to compress PDF output on RTD (#405)
Post-build compression is primarily used because it is
straightforward. It is also aware of the width of the images and as
such can be DPI-based. Even if we compressed the images before PDF
creation, we would likely still have some room for (minimal)
improvement in the rest of the PDF.