Bump dependencies (#441)
The only dependencies not bumped are those without new versions.

This means that in the future, we can add new packages that depend on
newer versions of Sphinx. It also likely puts us in a better position
regarding https://about.readthedocs.com/blog/2024/07/addons-by-default/.

This also bumps the minimum Python version 3.10, as that is the minimum
version all of our dependencies will allow.